requirements.txt
Django==4.2.9
djangorestframework==3.14.0
ccxt==4.1.87
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
